Continued refactoring.
