+++ CMakeLists.txt
@@ -21,7 +21,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
 # arma ex not required
-option(BUILD_ARMA_EX "Build armadillo ex" OFF)
+option(BUILD_ARMA_EX "Build armadillo ex" ON)
 
 # ClassifierContext::Context creation
 option(BUILD_CONTEXT_BUILDER "Build context builder" ON)
@@ -51,7 +51,7 @@ option(BUILD_INC_PREDICT "Build incrmental predictor" ON)
 option(BUILD_STEPWISE_PREDICT "Build stepwise predictor" ON)
 
 # build titanic example
-option(BUILD_TITANIC "Build titanic ex" OFF)
+option(BUILD_TITANIC "Build titanic ex" ON)
 
 # build multiclass pmlb driver
 option(BUILD_MULTICLASS "Build multiclass inductive boost classifier" OFF)
@@ -123,37 +123,58 @@ endif()
 
 include_directories(${ARMADILLO_INCLUDE_DIRS})
 
+################
+# OBJECT FILES #
+################
+# object files
+add_library(DP OBJECT DP.cpp)
 add_library(loss OBJECT loss.cpp)
-add_library(score2 OBJECT score2.cpp)
-add_library(utils OBJECT utils.cpp)
+add_library(pmlb OBJECT pmlb.cpp)
+add_library(gradientboostclassifier OBJECT gradientboostclassifier.cpp)
 add_library(replay OBJECT replay.cpp)
+add_library(utils OBJECT utils.cpp)
+add_library(score2 OBJECT score2.cpp)
 add_library(classifier OBJECT classifier.cpp)
 add_library(regressor OBJECT regressor.cpp)
-add_library(DP OBJECT DP.cpp)
+add_library(compositeclassifier OBJECT compositeclassifier.cpp)
 add_library(threadsafequeue OBJECT threadsafequeue.cpp)
-add_library(gradientboostclassifier OBJECT gradientboostclassifier.cpp)
 # add_library(gradientboostmulticlassifier OBJECT gradientboostmulticlassifier.cpp)
+
+###########
+# LINKAGE #
+###########
+target_link_libraries(DP PUBLIC utils score2
+	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 
 target_link_libraries(loss PUBLIC $<IF:$<BOOL:${USE_AUTODIFF}>,autodiff::autodiff,> 
 	${ARMADILLO_LIBRARIES} 
 	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES}
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 
-target_link_libraries(classifier PUBLIC utils ${ARMADILLO_LIBRARIES}
-	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES}
-	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)	
-
-target_link_libraries(regressor PUBLIC utils ${ARMADILLO_LIBRARIES}
-	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES}
-	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)	
-
-target_link_libraries(DP PUBLIC utils 
+target_link_libraries(replay PUBLIC utils threadsafequeue gradientboostclassifier
+	${ARMADILLO_LIBRARIES}
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
+
+target_link_libraries(utils PUBLIC loss
+	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
+
+target_link_libraries(compositeclassifier PUBLIC utils score2 classifier DP
+	${ARMADILLO_LIBRARIES}
+	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES}
+	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)	
+
+target_link_libraries(classifier PUBLIC ${ARMADILLO_LIBRARIES}
+	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES}
+	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)	
+
+target_link_libraries(regressor PUBLIC ${ARMADILLO_LIBRARIES}
+	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES}
+	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)	
 
 target_link_libraries(threadsafequeue PUBLIC pthread 
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 
-target_link_libraries(gradientboostclassifier PUBLIC utils classifier DP 
+target_link_libraries(gradientboostclassifier PUBLIC utils loss score2 DP classifier compositeclassifier
 	${ARMADILLO_LIBRARIES} 
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 
@@ -161,6 +182,10 @@ target_link_libraries(gradientboostclassifier PUBLIC utils classifier DP
 # 	${ARMADILLO_LIBRARIES} pthread 
 # 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 
+
+###############
+# EXECUTABLES #
+###############
 # Armadillo example
 if (BUILD_ARMA_EX)
   add_executable(arma_ex arma_ex.cpp)
@@ -172,21 +197,22 @@ endif()
 # Context creation
 if (BUILD_CONTEXT_BUILDER)
   add_executable(createContext createContext.cpp)
-  target_link_libraries(createContext LINK_PUBLIC ${Boost_LIBRARIES} utils 
+  target_link_libraries(createContext LINK_PUBLIC ${Boost_LIBRARIES} utils loss
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 endif()
 
 # DP solver example	
 if (BUILD_DP_SOLVER_EX)
   add_executable(DP_solver_ex DP_solver_ex.cpp)
-  target_link_libraries(DP_solver_ex DP pthread 
+  target_link_libraries(DP_solver_ex DP score2 pthread 
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 endif()
 		  
 # InductiveBoostClassifier driver on pmlb data
 if (BUILD_SINGLECLASS)
   add_executable(single_class_pmlb_driver single_class_pmlb_driver.cpp)
-  target_link_libraries(single_class_pmlb_driver loss utils gradientboostclassifier 
+  target_link_libraries(single_class_pmlb_driver loss utils
+	gradientboostclassifier
 	$<IF:$<BOOL:${USE_AUTODIFF}>,autodiff::autodiff,> ${ARMADILLO_LIBRARIES} 
 	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES} 
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
@@ -195,7 +221,7 @@ endif()
 # InductiveBoostRegressor driver on pmlb data
 if (BUILD_REGRESSOR)
   add_Executable(regressor_pmlb_driver regressor_pmlb_driver.cpp)
-  target_link_libraries(regressor_pmlb_driver loss utils gradientboostclassifier
+  target_link_libraries(regressor_pmlb_driver utils loss gradientboostclassifier
 	$<IF:$<BOOL:${USE_AUTODIFF}>,autodiff::autodiff,> ${ARMADILLO_LIBRARIES}
 	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES} 
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)	
@@ -227,7 +253,7 @@ endif()
 if (BUILD_STEPWISE_PREDICT)
   add_executable(stepwise_predict stepwise_predict.cpp)
   target_link_libraries(stepwise_predict LINK_PUBLIC ${Boost_LIBRARIES} loss utils 
-	gradientboostclassifier 
+	gradientboostclassifier replay
 	$<IF:$<BOOL:${USE_AUTODIFF}>,autodiff::autodiff,> ${ARMADILLO_LIBRARIES} 
 	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> 
 	${BLAS_LIBRARIES} $<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
@@ -237,8 +263,8 @@ endif()
 if (BUILD_TITANIC)
   add_executable(titanic_driver titanic_driver.cpp)
   target_link_libraries(titanic_driver LINK_PUBLIC ${Boost_LIBRARIES} 
-	${Boost_SYSTEM_LIBRARY} loss utils 
-	gradientboostclassifier 
+	${Boost_SYSTEM_LIBRARY} pmlb loss utils 
+	gradientboostclassifier replay
 	$<IF:$<BOOL:${USE_AUTODIFF}>,autodiff::autodiff,> ${ARMADILLO_LIBRARIES} 
 	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES} 
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
@@ -247,7 +273,7 @@ endif()
 # Multiclass classification driver on pmlb data
 if (BUILD_MULTICLASS)
   add_executable(multiple_class_pmlb_driver multiple_class_pmlb_driver.cpp)
-  target_link_libraries(multiple_class_pmlb_driver loss utils gradientboostclassifier 
+  target_link_libraries(multiple_class_pmlb_driver pmlb loss utils gradientboostclassifier 
 	gradientboostmulticlassifier 
 	$<IF:$<BOOL:${USE_AUTODIFF}>,autodiff::autodiff,> ${ARMADILLO_LIBRARIES} 
 	$<IF:$<BOOL:${USE_OPENMP}>,"${OpenMP_CXX_FLAGS}",> ${BLAS_LIBRARIES} 
@@ -258,7 +284,7 @@ if (BUILD_TESTS)
   set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/tests)
   add_executable(gtest_all gtest_all.cpp)
   target_link_libraries(gtest_all LINK_PUBLIC ${Boost_SYSTEM_LIBRARY} utils replay DP 
-	gradientboostclassifier ${GTEST_LIBRARIES} 
+	score2 gradientboostclassifier ${GTEST_LIBRARIES} 
 	${ARMADILLO_LIBRARIES} pthread 
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 endif()
@@ -267,7 +293,7 @@ if (BUILD_BENCHMARKS)
   set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/benchmarks)
   add_executable(benchmarks benchmarks.cpp)
   target_link_libraries(benchmarks benchmark::benchmark pthread utils
-	gradientboostclassifier 
+	score2 gradientboostclassifier 
 	${ARMADILLO_LIBRARIES} 
 	$<IF:$<BOOL:${GPROFILE_RESULTS}>,profiler,>)
 endif()
