+++ classifiers.hpp
@@ -1,15 +1,44 @@
 #ifndef __CLASSIFIERS_HPP__
 #define __CLASSIFIERS_HPP__
 
-#include "classifier.hpp"
-#include "model_traits.hpp"
+#include <utility>
 
-using namespace Model_Traits;
+#include <mlpack/core.hpp>
+#include <mlpack/methods/decision_tree/decision_tree.hpp>
+#include <mlpack/methods/decision_tree/decision_tree_regressor.hpp>
+#include <mlpack/methods/decision_tree/information_gain.hpp>
+#include <mlpack/methods/decision_tree/gini_gain.hpp>
+#include <mlpack/methods/decision_tree/random_dimension_select.hpp>
+#include <mlpack/methods/decision_tree/multiple_random_dimension_select.hpp>
+#include <mlpack/methods/random_forest/random_forest.hpp>
+
+#include "classifier.hpp"
+
+using namespace mlpack;
+using namespace mlpack::tree;
+using namespace mlpack::data;
+using namespace mlpack::util;
+
+namespace Model_Traits {
+
+  namespace ClassifierTypes {
+    using RandomForestClassifierType = RandomForest<>;
+    using DecisionTreeClassifierType = DecisionTree<>;
+
+    // using DecisionTreeClassifierType = DecisionTree<GiniGain, BestBinaryNumericSplit>;
+    // using DecisionTreeClassifierType = DecisionTree<GiniGain, BestBinaryNumericSplit, AllCategoricalSplit, AllDimensionSelect, true>;
+    // using DecisionTreeClassifierType = DecisionTreeRegressor<MADGain>;
+    // using DecisionTreeClassifierType = DecisionTreeRegressor<>;
+    // using DecisionTreeClassifierType = DecisionTreeRegressor<MSEGain, BestBinaryNumericSplit, AllCategoricalSplit, AllDimensionSelect, true>;
+    // using DecisionTreeClassifierType = DecisionTreeRegressor<InformationGain, BestBinaryNumericSplit, AllCategoricalSplit, AllDimensionSelect, true>;
+  
+  };
+} // namespace Model_Traits
 
 template<typename... Args>
 class RandomForestClassifierBase : 
   public DiscreteClassifierBase<double,
-				ClassifierTypes::RandomForestClassifierType,
+				Model_Traits::ClassifierTypes::RandomForestClassifierType,
 				Args...> {
 public:
   RandomForestClassifierBase() = default;
@@ -17,7 +46,7 @@ public:
   RandomForestClassifierBase(const mat& dataset,
 			     rowvec& labels,
 			     Args&&... args) :
-    DiscreteClassifierBase<double, ClassifierTypes::RandomForestClassifierType, Args...>(dataset, labels, std::forward<Args>(args)...) {}
+    DiscreteClassifierBase<double, Model_Traits::ClassifierTypes::RandomForestClassifierType, Args...>(dataset, labels, std::forward<Args>(args)...) {}
   
 };
 
@@ -42,7 +71,7 @@ public:
 template<typename... Args>
 class DecisionTreeClassifierBase : 
   public DiscreteClassifierBase<double,
-				ClassifierTypes::DecisionTreeClassifierType,
+				Model_Traits::ClassifierTypes::DecisionTreeClassifierType,
 				Args...> {
 public:
   DecisionTreeClassifierBase() = default;
@@ -50,7 +79,7 @@ public:
   DecisionTreeClassifierBase(const mat& dataset,
 			     rowvec& labels,
 			     Args&&... args) :
-    DiscreteClassifierBase<double, ClassifierTypes::DecisionTreeClassifierType, Args...>(dataset, labels, std::forward<Args>(args)...) {}
+    DiscreteClassifierBase<double, Model_Traits::ClassifierTypes::DecisionTreeClassifierType, Args...>(dataset, labels, std::forward<Args>(args)...) {}
 };
 
 class DecisionTreeClassifier : 
@@ -74,5 +103,31 @@ public:
   {}
 };
 
+namespace Model_Traits {
+
+  template<typename T>
+  struct model_traits {
+    using datatype = double;
+    using integrallabeltype = std::size_t;
+    using model = ClassifierTypes::DecisionTreeClassifierType;
+    using modelArgs = std::tuple<std::size_t, std::size_t, double, std::size_t>;
+  };
+
+  template<>
+  struct model_traits<DecisionTreeClassifier> {
+    using datatype = double;
+    using integrallabeltype = std::size_t;
+    using model = ClassifierTypes::DecisionTreeClassifierType;
+    using modelArgs = std::tuple<std::size_t, std::size_t, double, std::size_t>;
+  };
+
+  template<>
+  struct model_traits<RandomForestClassifier> {
+    using datatype = double;
+    using integrallabeltype = std::size_t;
+    using model = ClassifierTypes::RandomForestClassifierType;
+    using modelArgs = std::tuple<std::size_t, std::size_t, std::size_t>;
+  };
+} // namespace Model_Traits
 
 #endif
