+++ compositeclassifier.cpp
@@ -1,0 +1,1 @@
+#include "compositeclassifier.hpp"
