+++ incremental_driver.hpp
@@ -14,7 +14,6 @@
 
 #include "loss.hpp"
 #include "gradientboostclassifier.hpp"
-#include "replay.hpp"
 #include "utils.hpp"
